magit-bisecting-p: don't check bisect "status"
Previously `magit-bisectin-p' not only checked whether bisecting has
already started, but also whether it had not yet finished (found the
bad commit).  It also features an argument to NOT do the latter.

This was done for some commands that are not useful when bisecting is
already finished.  However they also do no HARM when run at that point
so we don't need that extra test, it only complicates things (like
subsequent refactorings).