magit-insert-status-local-line: don't show bisect information
That information should be displayed separately and the way the
information is currently being cached also is rather convoluted
and leaky.  So we remove this first (it is not essential) and
then, in a later commit, add a more suitable substitute.

Remove the function `magit--bisect-info-for-status' which was
used to format the bisect part of the string displayed by
`magit-insert-status-local-line'.