sanely cache the output of the last bisect command
Both the old and the new implementations dealt with this problem:
`git bisect CMD' does print some useful information, which cannot
be fully retrieved from the .git/BISECT_* files it also generates.

We could reimplement a lot of the `git-bisect' script and `bisect.c' but
that would be overkill.  In most cases the information isn't that
important; but still quite nice to have.  One case where we cannot get
to it is when the user runs `git bisect CMD' from the command line.  In
that case the information is in the shell, and we just cannot get to it
- so be it, the user probably saw it there.

--

This was the setting.
Now comes how the old implementation tried to deal with it:

`magit-run-git-bisect' (formerly known as `magit--bisect-cmd') used
`magit-git-lines' (should have used `magit-git-line') to get the first
line, and stored it in a buffer-local variable in the status buffer.

That's problematic because the user might delete that buffer at any time
in which case we would lose the cached information.  Also the buffer
might not exist in the first place, resulting in an error.

Then it went on throwing everything but the first line away.  In most
cases that is acceptable, but not when an error occurs or when using the
`run' command.

Also because `magit-git-lines' is used instead of some function that
ultimately uses `magit-run*' there is no error handling whatsoever.  If
something goes wrong the user has no choice but to rerun the whole
bisect on the command line, just to find out what went wrong.

Then, probably to deal with some of the above issues, it went on to
reimplement `magit-run*', including its own filter and sentinel, just
for the sake of `magit-bisect-run'.

--

This is what the new implementation does:  Use `magit-run-git*' (which
ultimately uses `magit-run*') like for all other user commands.  Git's
output is placed in the process buffer as usual.  The output is *also*,
and that's the *only* customization actually required, inserted into
the file .git/BISECT_CMD_OUTPUT (using a special filter function).
That's it.

That file is later used when displaying information in the status
buffer. The next commit will actually add the new "inserter" function
responsible for doing so.