magit-insert-bisect-output: new function
This uses the new "command output cache" implemented in the previous
commit and replaces `magit--bisect-info-for-status', which was removed
in the commit before that (because it used the old, broken caching
mechanism).

We use a "real" section with a heading and a body, instead of a "line"
section, because while the first line usually contains the interesting
part, that is not always the case.  The body is hidden by default.

The cached output is inserted verbatim in most cases, except when we
*know* the first line does not contain the "result".  Currently this the
only covered special case is when the `run` command was run.  In that
case the last line (mentioning the "first bad commit") is copied to the
front.  Other special cases can be added as they surface.

If there is no cached information, inform the user about why that is so,
instead of throwing the arms in the air, as the old implementation did.