magit-bisect-run: sanitize
The command `magit-bisect-run' is a wrapper around `git bisect run CMD
[ARGS...]', obviously.  All it has to do is to read `CMD [ARGS...]' from
the user and then pass that on to `git bisect run'.

But until now, after reading the input, it instead stored the user input
in a file, made it executable, called `git bisect run' with that file as
CMD, and then later and somewhere else (this was removed in the previous
commit) removed the file again.  Well, we don't need to do any of that...

Also we now read user input the same way `shell-command' does - if it's
good enough for that, then it is good enough for us too.