magit-read-rev-range: return atom if no range was read
This function is used to read a range *or a single rev*.  Changes
elsewhere assumed this function returns an atom when a single rev
was read, but until now that wasn't the case.

That broke the "ranged" log when it was used to show the history
of a branch (or any rev) different from the current branch.

Yes, there is a lot wrong here, this commit only fixes a bug that
resulted from doing some of the preparation necessary to fix the
bigger issues, not carefully enough.  Fixing the real issues still
require additional preparation.