remove doc-strings of section creation functions
Remove the doc-strings of `magit-new-section', `magit-cancel-section',
`magit-with-section', `magit-insert-section', `magit-git-section',
`magit-create-buffer-sections', and `magit-propertize-section'.  Of
these only `magit-with-section' will survive the subsequent refactoring.

New doc-strings will be added once the current refactoring is over.

Keeping these doc-strings, which in some cases are insufficient, would
make refactoring harder than it need be: changes would have to be
written trice: once in Lisp and twice in English.