remove introductory comment about sections
It's a noble goal to provide some documentation for those using the
source, but since this was the only such comment and because it was
also not informative enough, it did not really do any good.

I am all for literate programming, but that has to wait until after
the cleanup is over and after the next release.  And even then we
might not do it.