don't try to be smart about not highlighting sections
A major problem with the section implementation is that there is next to
no consistency regarding the meaning of the slots TITLE, TYPE, and INFO.
In different contexts and for different "kinds" (let's not use "type" to
avoid confusion :-) of sections are used to "identify" (what that means
also isn't consistent) sections.

As a result that code written with one use-case in mind cannot be used
elsewhere, leading to duplication and a growing number of inconsistencies.

One anomaly with likely is not only the result of this confusion but
probably the cause of many of these inconsistencies is that "a section
not having a title" is the (only) way of creating a section that is
never highlighted.

Instead add new slot HIGHLIGHT with defaults to t.