give calls to magit-git-section a uniform appearance
As much as possible format all calls to `magit-git-section' the same.

If possible put the TYPE and BUFFER-TITLE arguments on the same line
as `magit-git-section'.  When BUFFER-TITLE is calculated "inside" the
call to `magit-git-section' and that takes more than one line or a very
long line then put it on separate line(s).  Always put WASHER on its own
line; if it is a symbol quote with #' not just ', providing semantic
information to the reader.

I am doing this now because it helps looking for inconsistencies that go
beyond appearance.