identify branch-manager sections using symbols
In `magit-refresh-branch-manager' use the symbol `branchbuf' instead
of the string "branches" as the TYPE of the inserted top section.

In `magit-wash-branches' use the symbol `local' as the TYPE and so
called TITLE of the respective section.  Previously the string "local"
was used as TYPE and nil was used as TITLE.