identify all buffer sections, using symbol ending in `buf'
In `magit-refresh-reflog-buffer use `reflogbuf' instead of `reflog' as
the TYPE of the inserted top section. In `magit-refresh-wazzup-buffer'
use `wazzupbuf' instead of `wazzup'.

Now all such top-level buffer sections have a symbol as TYPE and the
symbol names of all of them end with "buf" (with no separating dash,
mostly for historic reasons).

Actually this is not quiet true.  The `logbuf' and `reflogbuf' section
are wrapped in an additional, otherwise empty, and unnecessary `log'
section.  But I noticed that to late to fix it in this commit.