magit-with-section: locally bind section to named variable
This prepares for the eventual removal of the troublesome variable
`magit-top-section'.  Three of the difficulties resulting from the
use of that variable are hinted at here, but there are more.

Add a new argument SECTION to the ARGLIST of `magit-with-section'.
It has to be a symbol and it is bound to the section being created
by `magit-with-section'.

Something similar is done in `magit-section-{action,case}'.

Do not actually make us of the new binding yet.  Except in one case
where not doing so would result in a surprising breakage (diffs can
no longer be expanded).  The main change in this commit is related,
but one would not expect this to break `magit-wash-raw-diff' - but
being predictable is not exactly a property of `magit-top-section'.

It might seem that there are not many uses of `magit-with-section'
that could actually make use of SECTION, because there are not many
that currently use `magit-top-section'.  That however is an illusion
- there are many sections that use it indirectly by calling functions
such as `magit-set-section-info' which do manipulate the section
stored in `magit-top-section', when their optional SECTION argument
is nil (which it always is).  These functions were presumably
created in the first place because having uses of `magit-top-section'
appear everywhere was troubling (or at least noisy) and someone
wanted to hide that fact (effectively obscuring `magit-top-section'
even more).