magit-set-section-info: remove function
Remove `magit-set-section-info' and the obfuscation it caused; and
instead use `setf' with the "current section"s INFO slot as PLACE.
Since the previous commit, that section is commonly available as the
value of a known symbol (with limited dynamic scope).

In some other cases where the call to `magit-set-section-info' was not
lexically within a call to `magit-with-section' the section now has to
be passed using a function argument (from the function which uses the
`magit-with-section' macro, to the function which wants to manipulate
the sections INFO slot).  That's a good thing.