magit-make-commit-button: absorb magit-set-section
Extend `magit-make-commit-button' so that it not only turns a text area
into a button but at the same time turns it into a magit section.  The
second task was previously performed by `magit-set-section'.  The latter
is removed in this commit.  And so is the noisy code that was previously
required around the calls to these functions - they only ever appeared
in combination.

Now `magit-make-commit-button' is much more useful because it does
everything related to creating a commit button, and could therefor
potentially be used elsewhere.  Actually we should consider whether the
"button as a section" concept could be used for other things besides
commits.

Regarding the removed `magit-set-section' I have to say this: It is a
colosal failure if a function is advertised as universally useful (here
by its name and placement) even though it is not, AND only has a single
caller (`magit-wash-commit'), FOR WHICH it is COMPLETELY inappropriate,
because it only serves to COMPLICATE that one caller (for which it
presumably was written), AND there already is an existing abstraction
(`magit-make-commit-button') which only has to be modified SLIGHTLY to
ACTUALLY make it easier for the caller.