magit-create-log-buffer-sections: remove macro
Remove macro `magit-create-log-buffer-sections'.  Dong so leads to a
tiny bit of code duplication.  Actually because it was only used twice
this commit adds exactly as many lines as it removes.  It used to be
used more often but even then the existence of this macro was not
justified.

This macro is related to (and uses) `magit-create-buffer-sections'.  In
addition to what the latter does it also prepares for something else to
count how many log items were inserted and then inserts the "there are
more commits" button.  That is not really closely enough related to
the task of `magit-create-buffer-sections' (dealing with the failing
`magit-top-section'/magit-old-top-section' abstraction) that one could
easily guess what the removed macro does in addition to what the other
does.  So let's just put the code where one would expect it.