magit-insert-section: absorb magit-cancel-section
It's not used anywhere else and by providing this functionality as a
separate function we encourage its use, which we shouldn't do because
using it would likely be due to failure to understand the section api.
There is already enough abuse of that, let's make it harder for
contributors to shot themselves (and us) in the foot.