magit-insert-section: don't remove empty buffer sections
Don't remove empty buffer sections, but keep doing it for other
sections, e.g. for "Staged changes" when there are not actually any
staged changes.  But do not remove empty buffer sections.  They are
not actually empty, but usually contain the header, which provides
information that is relevant when a user tries to determine why git
did not produce any input.  Previously the section was removed and
replaced with one that just said "(empty)", which is not helpful at
all.  E.g. "Changes in HEAD..HEAD\n(empty)" is much more useful than
"(empty)"; it could help the user realize that he did not actually
type "HEAD~.." as intended but "HEAD..".

Move all the code related to this into `magit-insert-section',
previously some of it was inside `magit-create-buffer-section'.