magit-section struct: define CONTENT-BEGINNING as a real slot
A function `magit-section-content-beginning' already existed but it just
assumed that the content begins one line after the beginning of the
section.  However that is not always correct; sections exist that have
no heading in which case the section and its content begin at the same
position.  In the future there might also sections with multiline
headings.

In `magit-with-section' CONTENT-BEGINNING is initially set to the same
value as BEGINNING.  Provided its BUFFER-TITLE argument is non-nil
`magit-insert-section' then adjust the slot's value.