make section headings consistent
Add a colon at the end of the heading of "Pending changes".

After this commit section headings are consistent.  The headings of all
toplevel buffer sections, that actually have a heading, DO NOT end with
a colon.  Those sections that don't have a heading don't do so because
they either start out with a "line" subsection.  Or because their
content begins with something that serves well as a heading, and making
that actually a heading would cause complications but no significant
benefit.

The headings of non-buffer, non-line sections, that actually do have a
heading, DO end with a colon.

I would actually like this to be customizable using two distinct format
strings, but that can wait.  It would also be nice if the user could
choose whether the newline should be propertized as part of the heading,
which becomes relevant when the background color is set.