make section TITLE slot consistent
In the "more commits" sections created in `magit-refresh-log-buffer' and
`magit-refresh-reflog-buffer' store the same information in the struct
slot TITLE that is already stored in TYPE.  That obviously is wrong but
at least it now is consistently wrong across Magit.  The reason why we
cannot just drop that yet, is that in some places TYPE is used to
retrieve this information, while in others TITLE is used.  It's also
worth noting that in a few other cases INFO is used instead or as yet
another alternative way of accessing this value.

In the section created in `magit-wash-remote-branches-group' store an
identification string in TITLE, instead of that string prefixed with
some irrelevant noise.

After this change all sections either store the same (constant) symbol
in TITLE and TYPE, or they store a symbol in TYPE and either nil (only
buffer sections) or some information in the TITLE slot, which can be
used to uniquely identify the section and/or is necessary when
performing some action on the section.  That value usually is some
string, but it may also be a symbol, it is however in all cases an atom.

The actual TITLEs being changed here are:
* 2x  "longer"                        =>  'longer
* 1x  (concat "remote:" remote-name)  =>  remote-name