magit-with-section: absorb magit-new-section
Until some commits ago `magit-new-section' was used in one other place
but `magit-with-section', but that was a mistake and has been fixed.
Now we can merge `magit-new-section' into `magit-with-section' to
prevent contributors from bypassing the additional, and required, setup
in `magit-with-section'.

Additionally this allows tiding up the code a bit, but we don't do that
in this commit.  After this commit, everything happens in the exact same
order as before, even when that means doing some things, only to
override them later.  Besides the removal of `magit-new-section' there
is one exception: the `let' which binds `magit-top-section' wraps around
less code.