magit-insert-wazzup-commits: don't calculate when hidden
In `magit-insert-wazzup-commits' stop consulting `magit-old-top-section'
directly and instead rely on the value of the HIDDEN slot of the section
being inserted, as setup by `magit-with-section'.

This not only saves code and prepares for the removal of said variable,
it also gives a considerable speedup.  Previously a section, after being
expanded by the user would get recalculated on every refresh, even when
the user collapsed it again.  The result of that was that refreshing the
wazzup buffer got slower and slower, which was surprising to users who
took care to only ever have a few sections expanded.