magit-with-section: take over heading insertion
Move the code responsible for inserting the heading from
`magit-insert-section' to `magit-with-section' and adapt it accordingly.

Now callers of `magit-with-section' can get the same benefit regarding
headers, as those of `magit-insert-section' and `magit-git-section' do.
We don't actually use it with `magit-with-section' yet and how exactly
headers are inserted is also still subject to change.