magit-insert-section: properly cancel empty sections
While inserting a section using `magit-insert-section' it can turn out
that there is no output from the PROGRAM that could actually be washed
to create the section.  This is the case because the PROGRAM, usually
git, isn't run *before* calling `magit-insert-section', but as part of
calling it.

As a result this function sometimes needs to "cancel" a section.  In
other words it has to undo all setup that has already been done. The
common setup is actually done by (the expansion of) `magit-with-section'.
After `magit-insert-section' is done with inserting and washing the
section text, `magit-with-section' finalizes the section by doing some
things common to all sections.

These final steps are unnecessary for sections that were canceled, and
some commits ago `magit-with-section' has been taught not to perform
them when its BODY set SECTION to nil.  Now finally we actually do set
it to nil.

Not that the "root" or "buffer" section is never canceled; if its body
is empty the string "(empty)" is inserted instead.