magit-old-top-section: don't abuse for speedup
The variable `magit-old-top-section' is used to preserve state during a
refresh.  As of now the only state that is preserved when refreshing, is
whether the section is expanded or not.  Because of the way state is
preserved, determining the previous state involves a bit of work, which
is useless for sections that do not have a "body".  For such sections
being hidden holds no meaning, as the heading is always displayed, even
when hidden.

Previously `magit-old-top-section' was let-bound to nil around some code
that created such sections, effectively disabling the code that
determined the previous visibility.  That was very confusing: without
reading the code that actually used the variable's value, it was
impossible to tell what binding the variable to nil was supposed to do.

Here we just remove these let-bindings.  A later commit will make
"getting the previous state" generally more efficient and as a result
make it unnecessary to add a hack to bypass doing so.  Actually I think
that the current code also is efficient enough, making the hack being
removed here, rather pointless in the first place.