magit-{git,cmd}-insert-section: new macros
Add two new macro that insert the washed output of a program as a
section: `magit-cmd-insert-section' and `magit-git-insert-section'.
These macros are convenience wrappers around `magit-with-section', just
like the functions `magit-insert-section' and `magit-git-section', which
they replace.  Unlike the latter they preserve most of the flexibility
of `magit-with-section'.

Members of the ARGS argument don't need to be strings.  If a member is
nil it is filtered out.  If a member is a list it is treated as if its
members were members of ARGS itself.  This is necessary because `apply'
can no longer be used (because we are dealing with macros).  But doing
it this way is also more convenient.