magit-with-section: set magit-section property
Set the `magit-section' text property in `magit-with-section'
instead of in `magit-propertize-section', which was called by
`magit-create-buffer-sections'.

Remove `magit-propertize-section', which now isn't used anymore.

The previous approach was to delay setting this property until all
sections have been inserted.  Then, starting at the root section,
recursively set the property for all sections.  This way the property
was first set to the parent section and then set again to the child
section.

Now, as soon as a section is inserted, the property is set, but only
for the part of its text that doesn't have it set yet, and therefor
does not belong to a child section.

Both approaches come with an (insignificant) performance penalty;
trade one for the other.  The approach we are now using is better
because the creation of a section should, as much as possible, be
"complete" when `magit-with-section' returns and because this is a
step towards the eventual removal of `magit-create-buffer-sections'.