sufficiently dynamic scope is indistinguishable from magic
In past commit messages I have been bitching a lot about the variable
`magit-top-section'; grep history for a good rant :-)

This commit finally replaces it with two distinct variables, each
taking responsibility for one, and only one, of its two remaining and
very distinct objectives.

One of these variables is the new `magit-root-section'.  It is defined
as an automatically buffer-local variable, and should only ever be used
as such.  Its value is the, one and only, "root" section of the buffer.
A buffer can only have one root section; all other sections have to be
descendants of that section.

The value of `magit-root-section' is set by `magit-with-section' when
that inserts the root section.  This macro automatically detects when
that is the case, the user just has to follow the conventions.  (Which
still need to be documented.  I might also later add sanity asserts.)
Outside of this macro the value of this variable should only ever be
read.

The old `magit-top-section' was also abused to represent the "section
that is the parent of the one currently being inserted".  It has also
been used to represent the "section currently being inserted", but that
was fixed some time ago.

So the value of `magit-with-section--parent', the other new variable
to replace `magit-top-section', is the parent of the section being
inserted, not the one being inserted.

Then there also was `magit-old-top-section' whose value is the section
that used to be the root section before the buffer was erased to refresh
it, by inserting a new generation of sections.  This variable is renamed
to `magit-with-section--oldroot'.

The symbol name of these two new variables contain the full name of
`magit-with-section' because they should never ever be used anywhere
else.  They are used to "pass" information from one instance to another.
Dynamic bindings have to be used for this purpose, because the
alternative would lead to unnecessary complications.  While passing
values like this is a bit ugly, all the trickery is now contained in
`magit-with-section'.

Essentially the magic now happens in this part of `magit-with-section':

  (let ((s (car arglist)))
    `(let ((,s (make-magit-section
                ...
                :parent magit-with-section-parent)))
       ...
       (let ((magit-with-section--parent ,s)
             (magit-with-section--oldroot
              (or magit-with-section--oldroot
                  (unless magit-with-section--parent
                    (prog1 magit-root-section
                      (setq magit-root-section ,s))))))
         ,@body)
       ...))

If you think that is bad, then imagine what it was like when this logic
was spread all across magit.el; `magit-with-section--parent', `s', and
`magit-root-section' all had the same symbol name; that variable was
used for yet some other purposes; and much of the codebase was in denial
by hiding its uses inside various "getter" functions.