magit-with-section: set invisible property
Set the `invisible' property in `magit-create-buffer-sections'.

The creation of a section should, as much as possible, be "complete"
when `magit-with-section' returns.  Unfortunately this commit isn't
enough for that to come true.  The code that does set the property
is still only run once we are done creating the root section.

To change that some prior changes to the "section visibility" code
would necessary.  This will be taken care of another time.

But this commit is still beneficial is a step towards the removal of
`magit-create-buffer-sections'.