magit-create-buffer-sections: remove macro
Remove the macro `magit-create-buffer-sections', after moving its last
responsibilities elsewhere: inhibit read-only in `magit-with-section'
and erase the buffer in `magit-mode-refresh-buffer'.  Everything else
it used to do has already been moved in earlier commits, mostly to
`magit-with-section', though.

This theoretically makes it possible to use most functions that insert
one section (or many sections that are all children of exactly one
section) as a "buffer refresh" function.  In practice a few more
preparations are still necessary.