magit-status-line-align-to: move definition
Move the definition of the variable `magit-status-line-align-to' into
the same section as the definitions of custom options, but don't turn it
into an option, just yet.  It will likely be renamed at some later time,
and doing so causes less problems if it is a variable people do not know
about.  We move it now because so that, after the next commit, is still
defined before its first use.