magit-insert-line-section: new macro
Replace the function `magit-insert-status-line' with the new macro
`magit-insert-line-section'.  The latter uses `magit-with-section',
so the inserted text is now part of a section.

We don't actually make use of the header lines now being sections
yet; the type of all such sections is currently the symbol `line'
and the section info is set to nil.  But that will change.