magit-wash-diff: emergency fix
Recent refactoring to the code responsible for creating sections did not
take into account that washing diffs is done in a way different from how
washing usually happens; inserting a diff involves washing git output
twice, and even though we are washing git output `magit-with-section' is
also.  That, unlike `magit-git-insert-section' does not cancel an empty
section and therefor returns t.  That led to an endless loop when for
empty diffs.

This commit is just a quick fix to avoid the loop.  A proper fix will
involve not washing twice, a change I intended to make anyway.