magit-wash-diff: properly cancel empty section
In `magit-wash-diff' properly cancel the section if it is empty by
setting SECTION to nil.  This still does not insert the "(empty)"
string, because that is implemented in `magit-cmd-insert-section',
not `magit-with-section'.  I am still undecided but I will probably
not move that feature and instead refactor diff washing so that it
is done in only one wash cycle (like most other things).