magit-grep: define as alias of vc-git-grep or lgrep
Remove the function definition of `magit-grep' and instead define it as
an alias of `vc-git-grep', or if `vc-git' isn't loaded (which it usually
is) then as an alias of `lgrep'.  The former derives from the later.

Using `git grep' without specifying a revision, like the old
`magit-grep' did and as `vc-git-grep' also does, does not gain anything
over just using `lgrep'.  Except that it can be used when grep and find
are not installed, of course.  What would be useful is if there was an

Eventually I will write a new implementation of `magit-grep'.  While
waiting for that to happen we use an alias instead of the old definition
because the latter was severely crippled, making it much *less* useful
than just plain `lgrep' or `rgrep'.