magit-checkout-branch-at-point: new command
Add the new command `magit-checkout-branch-at-point' and bind RET to it
in `magit-branch-manager-mode'.

Previously `magit-visit-item' did checkout the current branch when
invoked on a `branch' section, but that is wrong and this commit also
drops that case.  Using the buffer local keymap is the right thing to do
here - we should not abuse the `magit-section-action' functionality.  If
we always did what we were previously doing here, then section actions
would just become another way of implementing local keymaps (whether
that be buffer local keymaps or the `keymap' text property), and a
rather poor one at that.

Only actions that are fairly close to one another should ever be grouped
in a `magit-section-action' or `magit-section-case'.  This isn't the
case here.  Bringing up some information in some buffer about the thing
at point in the current buffer, as done in the other cases, is *very*
different from checking out a branch.