magit-show-item-or-scroll: new function
Add new function `magit-show-item-or-scroll' to decide on behalf of the
commands `magit-show-item-or-scroll-{up,down}' whether the commit buffer
should be scrolled and do it if appropriate.  Previously that was done
in `magit-show-{commit,stash}', which lead to these commands having an
interface that leaked implementation details.

Now that `magit-show-{commit,stash}' are not overloaded anymore, replace
their SELECT argument with NOSELECT, and remove the SCROLL argument.
Also change the argument order.

Remove most of `magit-show-commit's doc-string, some of that will come
back later.