disband "Stash Mode" section
The command `magit-show-stash' is just another variant of `magit-diff',
much like `magit-diff-working-tree', `magit-diff-unstaged', and
`magit-diff-staged'.  It currently differs from these command in that it
does not just convert some value into something else that is understood
by `magit-diff'.  Instead it calls `magit-mode-init' directly; but it
does use the same refresher, namely `magit-refresh-diff-buffer'.  So we
move `magit-show-stash' into the same section as `magit-diff' and the
other variants.  It will be renamed in a later commit.

The functions `magit-wash-stashes' and `magit-wash-stash' are not used
to display a list of stashes, as such they are only loosely related to
`magit-show-stash'.  We move them next to `magit-insert-stashes', which
is the only function that uses them.  In a later commit these functions
will be merged into one.