magit-set-buffer-margin: new function
Add function `magit-set-buffer-margin' as a smarter alternative to
`set-window-margin'.  Margins are local to a window, not a buffer.
In practice this is inconvenient.

This new function is similar to the built-in function in that it
arranges for the margin to be setup in the current window.  Additionally
it uses the `window-configuration-change-hook' to update the margin when
appropriate.  Effectively it gives the illusion that the margin is local
to a buffer.

Currently it always sets the right margin and leaves the left margin
alone; but that could be extended if the need arises.