magit-make-margin-overlay: new function
To put some text in the margin one has set the `display' property of
some text appearing on that line.  One should use an overlay for that
purpose, to avoid changing the appearance of the actual buffer text.
That's more or less how it works, actually it is more complicated.
Anyways one shouldn't have to worry about this, it's just how this is
implemented on the Emacs side.  This function hides the strangeness.