magit-with-emacsclient: delete leftover server
It is possible that a server file is not properly deleted after the
Emacs instance in which the server was running has terminated.  That
could happen if Emacs has crashes; there might be other reasons.

If the `server-name' is "server", then this isn't a big problem as we
then just append the pid of the process to the `server-name' and use
that instead.  If the server named "server" does indeed no longer exist,
that is not so nice, because we needlessly use a non-standard name.  But
we cannot do anything about it without jeopardizing performance (by
attempting to connect to the server, just to see whether that works).

If we cannot use "server" then we append the pid.  In the unlikely case
that the server file for "serverPID" already exists, we now
unconditionally force delete that server.  That server cannot possibly
be still in use because PID is unique and we already know it is not
running in the current instance.