magit-show: correctly deal with special symbol revs
Like other commands and internal functions `magit-show' still has
to deal with "special revs" represented by the symbols `index` and
`working`.  While I intend to remove that special case, we are not
there yet.  Before this commit `magit-show' did not properly deal
with these revs, which seemed like a small problem because there
are better ways of getting the same information in these cases.
But that of course does not mean it is impossible to use this
command under such conditions; so make the necessary adjustments.

Close #1083.