magit-refresh-commit-buffer: don't use --no-abbrev-commit
Don't use `--no-abbrev-commit' when calling `git log -1' in
`magit-refresh-commit-buffer'; it isn't supported by Git version
1.7.2.5, which we still want to support.  Also this option didn't
have quite the intended effect anyway: the parent hashes were still
abbreviated.  At some point we will have to deal with this again:
an uniquely abbreviated hash to stay unique.  But there are so many
other places that don't care about this, that it doesn't really make
a difference to have one more, for now that is