magit-stgit.el: remove concept of a marked commit
Supporting a "marked" commit is problematic because that for every
feature being implemented it has to be decided whether that acts on the
"top", "marked", or "at point" commit, provide a way to select the
other.  And then "the other" is yet again divided. Etc.  And then that
also differs between various features.  And then there might be
different opinions among users what is the right thing to do.  Etc.

Let's just not do it.  Also Magit's StGit is rather basic anyway, given
the few commands we actually have it is not necessary to have this
convenience feature (for users, for maintainers it's not a feature but a
nightmare).