magit-run-topgit{*,-async}: new functions
Considering how few callers these functions have this might be overkill.
On the other hand this makes `magit-run-topgit{*}' most consistent with
`magit-run-git{*}', and maybe someone wants to extend our topgit support
sometime, at which point this will come in handy.  (Some arguments of
`magit-run-git*' that are not actually ever used have been left out, so
this is not completely consistent, but for the better.)