magit-diff-range: new function (kludge)
The commands `magit-ediff' and `magit-show' badly need to be sanitized.
Unfortunately not only they are rather messy but also some data they
consume, the "diff range".  And then they way that value is determined
and stored is also rather wacky.  It's a chicken and egg problem, and
I have chosen to start here.

Eventually the new function `magit-diff-range' will be replaced by
something saner and more complete.